pinctrl: ingenic: Make dual-edge emulation resistant to bouncing
Reading the GPIO at each edge to determine what should be the next
trigger is a bad idea, because at the time the interrupt occurs, the
GPIO line may be bouncing.

Instead, when switching the trigger edge to emulate dual-edge handling,
cache the previous trigger level, and use the opposite one.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>